refactor ast route helpers
